app.py
# app.py
import streamlit as st
import io
from PIL import Image

# Set page configuration
st.set_page_config(layout="wide", page_title="Pixel Prediction")

# Header
st.title("Pixel Prediction")

# Create two columns: one for the image display and upload, one for the measurement tool
col1, col2 = st.columns([3, 1])

# Main content area (left column)
with col1:
    # Create a placeholder for the image display
    image_placeholder = st.empty()

    # File uploader
    uploaded_file = st.file_uploader("Upload an image", type=["jpg", "jpeg", "png"])

    if uploaded_file is not None:
        # Read the file and display the image
        image = Image.open(uploaded_file)
        image_placeholder.image(image, caption="Uploaded Image", use_column_width=True)

        # Placeholder for deforestation count (this would be calculated by your ML model)
        deforested_trees = 100  # This is a placeholder value
        st.write(f"Number of deforested trees: {deforested_trees}")

# Measurement tool (right column)
with col2:
    st.subheader("Deforestation in %")

    # Placeholder for deforestation percentage (this would be calculated by your ML model)
    deforestation_percentage = 25  # This is a placeholder value

    # Create a progress bar to represent the deforestation percentage
    st.progress(deforestation_percentage)
    st.write(f"Deforestation: {deforestation_percentage}%")

# Note about future functionality
st.info("Placeholder, Number of Trees deforested.")
