requirements.txt
absl-py==1.3.0
altair==4.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
branca==0.8.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.32
dacite==1.8.1
dask==2022.7.1
db-dtypes==1.0.4
deap==1.3.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
etils==0.8.0
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.12.4
flake8==4.0.1
flatbuffers==22.9.24
folium==0.18.0
fonttools==4.38.0
fsspec==2022.10.0
gast==0.4.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
google-trans-new==1.1.9
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.48.2
h11==0.12.0
h5py==3.7.0
htmlmin==0.1.12
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.16.4
idna==3.4
ImageHash==4.3.1
imageio==2.22.2
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.15.3
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.10
jsonschema==4.16.0
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.5.0
jupyter-resource-usage==0.6.3
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyter_latex_envs==1.4.6
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.16.1
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
libclang==14.0.6
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
memoized-property==1.0.3
missingno==0.5.1
mistune==0.8.4
mpmath==1.3.0
multimethod==1.8
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==6.5.4
nbformat==5.7.0
nbresult==0.1.0
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.7
notebook==6.4.12
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandas-gbq==0.17.9
pandas-profiling==3.3.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
pmdarima==2.0.1
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==5.28.3
psutil==5.9.3
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pydata-google-auth==1.4.0
pydeck==0.8.0b4
pyflakes==2.4.0
Pygments==2.13.0
pylint==2.14.5
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
pytest==7.1.3
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.3.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.99
setuptools-scm==6.4.2
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
statsmodels==0.13.2
stopit==1.1.2
streamlit==1.40.2
streamlit_folium==0.23.2
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
terminado==0.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tokenizers==0.14.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
toolz==0.12.0
torch @ https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.1.0%2Bcpu.cxx11.abi-cp310-cp310-linux_x86_64.whl#sha256=88f1ee550c6291af8d0417871fb7af84b86527d18bc02ac4249f07dcd84dda56
tornado==6.2
TPOT==0.11.7
tqdm==4.64.1
traitlets==5.5.0
transformers==4.34.0
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.12
validators==0.20.0
visions==0.7.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wordcloud==1.9.3
wrapt==1.14.1
xgboost==1.6.2
XlsxWriter==3.0.3
xyzservices==2024.9.0
yapf==0.32.0
ydata-profiling==4.2.0
zipp==3.10.0
